Triton_forecaster/flask_app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
curl_cffi==0.13.0
Flask==3.1.2
frozendict==2.4.6
frozenlist==1.7.0
gevent==25.8.2
geventhttpclient==2.0.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
multidict==6.6.4
multitasking==0.0.12
newsapi-python==0.2.7
nltk==3.9.1
numpy==1.24.4
pandas==2.3.2
pandas_ta==0.3.14b0
peewee==3.18.2
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-rapidjson==1.21
pytz==2025.2
regex==2025.9.1
requests==2.32.5
scikit-learn==1.7.1
scipy==1.11.4
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
tqdm==4.67.1
tritonclient==2.60.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
xgboost==1.6.2
yarl==1.20.1
yfinance==0.2.65
zope.event==5.1.1
zope.interface==7.2
matplotlib